requirementsProducer.txt
requests
confluent_kafka
pandas
cassandra-driver